// frontend/src/app/api/model/discoveredPort.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DiscoveredPort { 
    /**
     * Technical name of the port (ifDescr)
     */
    portNumber?: string;
    /**
     * OID (ifIndex) of the port
     */
    oid?: string;
}